# Backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Unora mobile application",
        "title": "Unora Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/Unora-app/Backend",
            "email": "support@unora.app"
        },
        "license": {
            "name": "Private",
            "url": "https://github.com/Unora-app/Backend"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/credit-packages": {
            "get": {
                "security": [
                    {
                        "AdminAPIKey": []
                    }
                ],
                "description": "Get all credit packages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List credit packages",
                "parameters": [
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include inactive",
                        "name": "includeInactive",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Credit packages",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/credit-packages/{packageId}": {
            "patch": {
                "security": [
                    {
                        "AdminAPIKey": []
                    }
                ],
                "description": "Update a credit package",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update credit package",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Package ID",
                        "name": "packageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Package updates",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_admin_dto.UpdateCreditPackageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Package updated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/dashboard": {
            "get": {
                "security": [
                    {
                        "AdminAPIKey": []
                    }
                ],
                "description": "Get aggregated dashboard statistics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get dashboard stats",
                "responses": {
                    "200": {
                        "description": "Dashboard stats",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_admin_dto.DashboardStatsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/hobby-options": {
            "get": {
                "security": [
                    {
                        "AdminAPIKey": []
                    }
                ],
                "description": "Get all hobby options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List hobby options",
                "parameters": [
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include inactive",
                        "name": "includeInactive",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hobby options",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AdminAPIKey": []
                    }
                ],
                "description": "Create a new hobby option",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create hobby option",
                "parameters": [
                    {
                        "description": "Hobby option details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_admin_dto.CreateHobbyOptionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Hobby option created",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/hobby-options/{optionId}": {
            "delete": {
                "security": [
                    {
                        "AdminAPIKey": []
                    }
                ],
                "description": "Soft delete a hobby option",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete hobby option",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hobby Option ID",
                        "name": "optionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hobby option deleted",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/reports": {
            "get": {
                "security": [
                    {
                        "AdminAPIKey": []
                    }
                ],
                "description": "Get paginated list of user reports",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List reports",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Report status filter",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reports",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_admin_dto.AdminReportListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/reports/{reportId}": {
            "get": {
                "security": [
                    {
                        "AdminAPIKey": []
                    }
                ],
                "description": "Get detailed report information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get report details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Report ID",
                        "name": "reportId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_admin_dto.AdminReportResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/reports/{reportId}/review": {
            "post": {
                "security": [
                    {
                        "AdminAPIKey": []
                    }
                ],
                "description": "Review a report and optionally take action",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Review report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Report ID",
                        "name": "reportId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Review details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_admin_dto.ReviewReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report reviewed",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "AdminAPIKey": []
                    }
                ],
                "description": "Get paginated list of users with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List users",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Account status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by email or name",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_admin_dto.AdminUserListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/users/{userId}": {
            "get": {
                "security": [
                    {
                        "AdminAPIKey": []
                    }
                ],
                "description": "Get detailed user information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get user details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_admin_dto.AdminUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AdminAPIKey": []
                    }
                ],
                "description": "Soft delete a user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_UnoraApp_be_internal_admin_dto.UpdateCreditPackageRequest": {
            "description": "Update credit package details",
            "type": "object",
            "properties": {
                "badgeText": {
                    "type": "string",
                    "example": "Popular"
                },
                "bonusCredits": {
                    "type": "integer",
                    "example": 5
                },
                "creditAmount": {
                    "type": "integer",
                    "example": 50
                },
                "description": {
                    "type": "string"
                },
                "discountPercent": {
                    "type": "number",
                    "example": 10
                },
                "isActive": {
                    "type": "boolean",
                    "example": true
                },
                "isPopular": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "Starter Pack"
                },
                "priceAmount": {
                    "type": "integer",
                    "example": 9900
                },
                "sortOrder": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "github_com_UnoraApp_be_pkg_response.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_UnoraApp_be_pkg_response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}